refactor: extract shared fetch-timeout-retry logic from duplicated Supabase client files
